fix: teacher preview plays word audio instead of example sentence audio (#710)
* fix: teacher preview plays word audio instead of example sentence audio

Teacher previews assignment 181 (vocab + reading mode). Items have
example_sentence text filled in but example_sentence_audio_url stored
as empty string ""; the preview audio plays the single-word audio
(item.audio_url) instead of the sentence audio the teacher expects.

Two compounding bugs:

1. ensure_example_sentence_audio / ensure_word_audio in preview_service:
   - In-memory check `not ci.example_sentence_audio_url` correctly
     treats "" as missing — items get queued for TTS generation
   - But the persistent UPDATE clause was only `WHERE ... IS NULL`, so
     when the column is "" (not NULL), rowcount=0 and the in-memory
     attribute is never set, leaving lazy TTS effectively a no-op
   - Fix: WHERE (col IS NULL OR col = '')

2. routers/teachers.get_assignment_preview never called the lazy TTS
   helpers at all. Even with bug 1 fixed, the teacher preview would
   still serve stale example_sentence_audio_url values because it
   skipped generation entirely.
   - Fix: mirror student endpoint pattern — collect vocab items across
     all contents, call ensure_example_sentence_audio for
     reading/rearrangement/word_cloze, ensure_word_audio for
     word_reading/word_spelling

Confirmed reproduction on staging assignment 181: 11 vocab items, all
with example_sentence text and example_sentence_audio_url="", audio
playback served item.audio_url (cached single-word TTS) for every
question.


* fix: address Claude review — move lazy TTS to live endpoint, drop dead code

Claude Code Review on PR #710 correctly identified that backend/routers/teachers.py
is shadowed by the backend/routers/teachers/ package. Production resolves
`from routers import teachers` to the package's __init__.py; the file is
never imported. The lazy TTS pre-pass added in this PR's first commit lived
at teachers.py:3878-3886 — dead code.

Confirmed by curl on staging: GET /api/teachers/assignments/181/preview/practice-words
returns 404 (route not registered). PR #709's preview_practice_words endpoint
also lived in the dead file and never deployed.

Changes:

1. routers/teachers/assignment_ops.py — the LIVE module:
   - Add lazy TTS pre-pass to get_assignment_preview() before
     build_assignment_preview() call. Mirrors the pattern in
     routers/students/assignments.get_assignment_activities. Mode-gated
     loads (only for vocab content types), idempotent helpers
     (ensure_example_sentence_audio / ensure_word_audio).
   - Add the preview_practice_words endpoint that PR #709 intended to
     ship — uses the existing _get_teacher_assignment helper and
     returns session_id=None, no PracticeSession row.
   - Module-level imports (not inline-conditional) per reviewer's
     style note.

2. routers/teachers.py — DEAD file:
   - Remove the lazy TTS pre-pass (this commit) and the
     preview_practice_words endpoint (from PR #709) that were never
     reachable in production.
   - Restore the original `from services.preview_service import
     get_sentence_fields` (without the dead _VOCABULARY_CONTENT_TYPES
     re-import).

3. services/preview_service.py — unchanged from first commit
   (`WHERE col IS NULL OR col = ''` fix is correct and applies live).

Verified live router registration:
    practice-words routes: ['/assignments/{assignment_id}/preview/practice-words']
    main preview route: ['/assignments/{assignment_id}/preview']

Out of scope per reviewer's "minor #3" note: the audio_by_id return value
from the lazy TTS helpers is still discarded (matching student endpoint
pattern). If the SQLAlchemy expire-on-commit silent-failure case ever
manifests, fix at all three call sites in one pass — not in this PR.


* fix: address Claude review round 2 — preview_practice_words gaps

Round-2 review on PR #710 flagged two ⚠️ items in the new
preview_practice_words endpoint (both pre-existing in the dead-code
version that was moved to the live router in commit fd4beb70):

1. Endpoint never lazy-generated word audio. With the
   `(IS NULL OR = '')` UPDATE-clause fix already in this PR, calling
   `ensure_word_audio` here makes the endpoint self-healing in the same
   way that `get_assignment_preview` is — a teacher previewing a vocab
   set with empty `audio_url` columns now gets TTS generated up-front
   instead of silent audio buttons.

2. SQL query had no content-type filter. For assignments mixing
   vocabulary sets with reading-passage content, non-vocabulary
   ContentItems (e.g. passage paragraphs) would leak into the words
   list. Added `Content.type.in_(_VOCABULARY_CONTENT_TYPES)` predicate
   — `_VOCABULARY_CONTENT_TYPES` is already imported at module level.

Both fixes are surgical and live in the same endpoint added in this PR.


---------
