+++ backend/routers/teachers.py
@@ -4727,75 +4727,3 @@ async def preview_word_selection_answer(
         "target_mastery": assignment.target_proficiency or 80,
         "achieved": False,
     }
-
-
-# ============ Sentence Making Practice Preview API ============
-
-
-@router.get("/assignments/{assignment_id}/preview/practice-words")
-async def preview_practice_words(
-    assignment_id: int,
-    current_teacher: Teacher = Depends(get_current_teacher),
-    db: Session = Depends(get_db),
-):
-    """
-    預覽模式專用：取得造句練習題目（艾賓浩斯記憶曲線系統的單字）
-
-    - 供老師預覽示範用，不建立 PracticeSession
-    - 不需要 StudentAssignment，直接從 Assignment 讀取
-    - 順序回傳前 10 個單字（沒有學生作答歷史可以排序）
-    - 回傳格式對齊 students 端 GET /assignments/{id}/practice-words，
-      但 session_id=None 表示預覽模式
-    """
-    # 確認老師擁有這個作業（支援 classroom_id 為 null 的即刻練習）
-    assignment = (
-        db.query(Assignment)
-        .filter(
-            Assignment.id == assignment_id,
-            Assignment.teacher_id == current_teacher.id,
-        )
-        .first()
-    )
-
-    if not assignment:
-        raise HTTPException(status_code=404, detail="Assignment not found")
-
-    # answer_mode 對齊學生端：dict / Enum / None 都要 normalise 成字串
-    answer_mode_value = assignment.answer_mode or "listening"
-    if not isinstance(answer_mode_value, str):
-        answer_mode_value = (
-            answer_mode_value.value
-            if hasattr(answer_mode_value, "value")
-            else "listening"
-        )
-
-    # 取得作業內所有 ContentItem（含 Content 以判斷類型，與其他 preview 一致）
-    content_items = (
-        db.query(ContentItem)
-        .join(Content)
-        .join(AssignmentContent)
-        .filter(AssignmentContent.assignment_id == assignment.id)
-        .order_by(ContentItem.order_index)
-        .limit(10)
-        .all()
-    )
-
-    words = [
-        {
-            "content_item_id": item.id,
-            "text": item.text or "",
-            "translation": item.translation or "",
-            "example_sentence": item.example_sentence or "",
-            "example_sentence_translation": item.example_sentence_translation or "",
-            "audio_url": item.audio_url or "",
-            "memory_strength": 0.0,
-            "priority_score": 0.0,
-        }
-        for item in content_items
-    ]
-
-    return {
-        "session_id": None,  # 預覽模式不建立 session
-        "answer_mode": answer_mode_value,
-        "words": words,
-    }
